sse-to-kafka/requirements.txt
sseclient
kafka-python
